# public/api-docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Dot.Analytics API",
        "description": "The Enterprise Intelligence Platform API for the Dot Ecosystem. Provides cross-platform intelligence, knowledge graph traversal, metric computation, and AI-powered recommendations.",
        "version": "v1",
        "contact": {
            "name": "Dot Ecosystem",
            "url": "https://infodot.app"
        },
        "license": {
            "name": "MIT"
        }
    },
    "servers": [
        {
            "url": "http://analytics.infodot.app/api",
            "description": "Current server"
        },
        {
            "url": "https://analytics.infodot.app/api",
            "description": "Production"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum API Token",
                "description": "Obtain a token via POST /api/v1/tokens/create or from the dashboard."
            }
        },
        "responses": {
            "Unauthorized": {
                "description": "Unauthenticated \u2014 missing or invalid API token."
            },
...
            "Unprocessable": {
                "description": "Validation error \u2014 check the errors field."
            },
            "TooManyRequests": {
                "description": "Rate limit exceeded. Retry after the X-RateLimit-Reset header value."
            }
        },
        "schemas": {
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/v1/intelligence/insights": {
            "get": {
                "tags": [
                    "Intelligence"
                ],
                "summary": "List cross-platform insights.",
                "operationId": "insights",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/Unprocessable"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "info",
                                "warning",
                                "critical"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        }
                    }
                ]
            }
        },
...
        "/v1/reports/{type}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate a JSON report. Types: insights, alerts, recommendations, metrics.",
                "operationId": "report-json",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/Unprocessable"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                },
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
...
}